checking

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: ftan <ftan@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/02 20:23:48 by ftan              #+#    #+#             */
-/*   Updated: 2021/10/03 01:21:30 by ftan             ###   ########.fr       */
+/*   Updated: 2021/10/03 01:28:24 by ftan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 int	ft_atoi(const char *str)
 {
 	size_t	i;
-	size_t	sign;
+	int		sign;
 	long	sum;
 
 	sign = 1;
